[TASK] Highlight metadata edit button with missing attributes
Highlighting of the metadata edit button in the filelist
list view was previously disabled as a hotfix. To provide
this highlighting again the LinkButton classes needed to
be xclassed due to the following reasons.

Adding a custom css class to the LinkButton is possible
and it gets appended to the css classes string which
results in the rendered a-tag looking something like
this: `<a class="btn btn-sm btn-default custom-class"
…`. This way the `btn-danger` css class is visible in the
rendered html in the backend. Due to the higher priority
of the `btn-default` css class styling the `btn-danger`
styling is overwritten and no highlighting is visible.
Therefore the highlighting must either be used on a
different html element or the `btn-default` css class
needs to be removed in this particular case. As adding
the highlighting to a different element was not possible
the LinkButton class had to be xclassed to replace the
`btn-default` css class whenever the specific
`required-attributes-missing` css class, previously
added by the FileListActionsEventListener, is found in
the LinkButtons html.

The isValid function needed to be overwritten as well,
as it compares with an instance of itsself when
validating the button.